Changes: * draft impl of cca ra - needs testing and some parts need to be coded - 90% done:   - avps impl   - pojos   - cca ra classes   - event fix
  JDiame sesssions need to be fixed = they require handle code for STR/STA ASR/ASA ACR/ACA and extension
* fixes to base
* adds avp utility class to perform some operations we always do

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/diameter@3991 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ core/jdiameter/impl/src/main/java/org/jdiameter/common/api/app/cca/IClientCCASessionContext.java
@@ -10,8 +10,8 @@ public interface IClientCCASessionContext {
 
 	void txTimerExpired(ClientCCASession session);
 	
-	int getDefaultCCFHValue();
-	int getDefaultDDFHValue();
+	long getDefaultCCFHValue();
+	long getDefaultDDFHValue();
 
 	void grantAccessOnDeliverFailure(ClientCCASession clientCCASessionImpl,Message request);
